Wip: Protected areas component
