+++ app/components/setup/protected-areas/styles.js
@@ -1,0 +1,7 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  map: {
+    ...StyleSheet.absoluteFillObject
+  }
+});
